chore: data backup context (#712)
