+++ src/renderer/Providers.tsx
@@ -2,48 +2,49 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 // General contexts.
-import { HelpProvider } from '@/renderer/contexts/common/Help';
-import { OverlayProvider } from '@/renderer/contexts/common/Overlay';
-import { TooltipProvider } from '@/renderer/contexts/common/Tooltip';
-import { ConnectionsProvider } from '@/renderer/contexts/common/Connections';
+import { HelpProvider } from '@app/contexts/common/Help';
+import { OverlayProvider } from '@app/contexts/common/Overlay';
+import { TooltipProvider } from '@app/contexts/common/Tooltip';
+import { ConnectionsProvider } from '@app/contexts/common/Connections';
 
 // Tabs contexts.
 import { TabsProvider } from '@/renderer/contexts/tabs/Tabs';
 
 // Main window contexts.
-import { AddressesProvider } from '@/renderer/contexts/main/Addresses';
-import { AppSettingsProvider } from '@/renderer/contexts/main/AppSettings';
+import { AddressesProvider } from '@app/contexts/main/Addresses';
+import { AppSettingsProvider } from '@app/contexts/main/AppSettings';
 import { BootstrappingProvider } from '@app/contexts/main/Bootstrapping';
-import { ChainsProvider } from '@/renderer/contexts/main/Chains';
-import { EventsProvider } from '@/renderer/contexts/main/Events';
-import { ManageProvider } from '@/renderer/contexts/main/Manage';
+import { ChainsProvider } from '@app/contexts/main/Chains';
+import { EventsProvider } from '@app/contexts/main/Events';
+import { ManageProvider } from '@app/contexts/main/Manage';
 import { SubscriptionsProvider } from '@app/contexts/main/Subscriptions';
-import { IntervalSubscriptionsProvider } from './contexts/main/IntervalSubscriptions';
-import { IntervalTasksManagerProvider } from './contexts/main/IntervalTasksManager';
+import { IntervalSubscriptionsProvider } from '@app/contexts/main/IntervalSubscriptions';
+import { IntervalTasksManagerProvider } from '@app/contexts/main/IntervalTasksManager';
+import { DataBackupProvider } from '@app/contexts/main/DataBackup';
 
 // Import window contexts.
 import { AccountStatusesProvider as ImportAccountStatusesProvider } from '@app/contexts/import/AccountStatuses';
 import { AddressesProvider as ImportAddressesProvider } from '@app/contexts/import/Addresses';
-import { ImportHandlerProvider } from './contexts/import/ImportHandler';
-import { AddHandlerProvider } from './contexts/import/AddHandler';
-import { RemoveHandlerProvider } from './contexts/import/RemoveHandler';
-import { DeleteHandlerProvider } from './contexts/import/DeleteHandler';
+import { ImportHandlerProvider } from '@app/contexts/import/ImportHandler';
+import { AddHandlerProvider } from '@app/contexts/import/AddHandler';
+import { RemoveHandlerProvider } from '@app/contexts/import/RemoveHandler';
+import { DeleteHandlerProvider } from '@app/contexts/import/DeleteHandler';
 
 // Settings window contexts.
-import { SettingFlagsProvider } from './contexts/settings/SettingFlags';
-import { WebsocketServerProvider } from './contexts/settings/WebsocketServer';
-import { WorkspacesProvider } from './contexts/settings/Workspaces';
+import { SettingFlagsProvider } from '@app/contexts/settings/SettingFlags';
+import { WebsocketServerProvider } from '@app/contexts/settings/WebsocketServer';
+import { WorkspacesProvider } from '@app/contexts/settings/Workspaces';
 
 // Actions window contexts.
-import { TxMetaProvider } from '@/renderer/contexts/action/TxMeta';
+import { TxMetaProvider } from '@app/contexts/action/TxMeta';
 
 // OpenGov window contexts.
-import { TracksProvider } from './contexts/openGov/Tracks';
-import { TreasuryProvider } from './contexts/openGov/Treasury';
-import { ReferendaProvider } from './contexts/openGov/Referenda';
-import { ReferendaSubscriptionsProvider } from './contexts/openGov/ReferendaSubscriptions';
-import { TaskHandlerProvider } from './contexts/openGov/TaskHandler';
-import { PolkassemblyProvider } from './contexts/openGov/Polkassembly';
+import { TracksProvider } from '@app/contexts/openGov/Tracks';
+import { TreasuryProvider } from '@app/contexts/openGov/Treasury';
+import { ReferendaProvider } from '@app/contexts/openGov/Referenda';
+import { ReferendaSubscriptionsProvider } from '@app/contexts/openGov/ReferendaSubscriptions';
+import { TaskHandlerProvider } from '@app/contexts/openGov/TaskHandler';
+import { PolkassemblyProvider } from '@app/contexts/openGov/Polkassembly';
 
 // Other imports.
 import { Theme } from './Theme';
@@ -71,7 +72,9 @@ const getProvidersForWindow = () => {
         IntervalTasksManagerProvider,
         EventsProvider,
         // Online status relies on other contexts being initialized.
-        BootstrappingProvider
+        BootstrappingProvider,
+        // Requires setting state from other contexts.
+        DataBackupProvider
       )(Theme);
     }
     case 'import': {
