+++ src/renderer/utils/ImportUtils.ts
@@ -6,22 +6,13 @@ import { Config as ConfigRenderer } from '@/config/processes/renderer';
 import { Flip, toast } from 'react-toastify';
 import type {
   AccountSource,
-  FlattenedAccounts,
   LedgerLocalAddress,
   LocalAddress,
 } from '@/types/accounts';
 import { getAddressChainId } from '../Utils';
-import { IntervalsController } from '@/controller/renderer/IntervalsController';
 import type { AnyData } from '@/types/misc';
 import type { ChainID } from '@/types/chains';
-import type { EventCallback } from '@/types/reporter';
-import type {
-  IntervalSubscription,
-  SubscriptionTask,
-} from '@/types/subscriptions';
 import type { IpcTask } from '@/types/communication';
-import { AccountsController } from '@/controller/renderer/AccountsController';
-import { SubscriptionsController } from '@/controller/renderer/SubscriptionsController';
 
 type ToastType = 'success' | 'error';
 
@@ -169,245 +160,11 @@ export const getSortedLocalLedgerAddresses = (
 };
 
 /**
- * @name importAddresses
- * @summary Extract address data from an imported text file and send to application.
- * (main renderer)
- */
-export const importAddresses = async (
-  serialized: string,
-  handleImportAddress: (ev: MessageEvent, fromBackup: boolean) => Promise<void>,
-  handleRemoveAddress: (ev: MessageEvent) => Promise<void>,
-  setAddresses: (a: FlattenedAccounts) => void
-) => {
-  const s_addresses = getFromBackupFile('addresses', serialized);
-  if (!s_addresses) {
-    return;
-  }
-
-  const p_array: [AccountSource, string][] = JSON.parse(s_addresses);
-  const p_map = new Map<AccountSource, string>(p_array);
-  const importWindowOpen = await window.myAPI.isViewOpen('import');
-
-  for (const [source, ser] of p_map.entries()) {
-    const parsed =
-      source === 'ledger'
-        ? (JSON.parse(ser) as LedgerLocalAddress[])
-        : (JSON.parse(ser) as LocalAddress[]);
-
-    // Check connection status and set isImported to `false` if app is offline.
-    const isOnline: boolean =
-      (await window.myAPI.sendConnectionTaskAsync({
-        action: 'connection:getStatus',
-        data: null,
-      })) || false;
-
-    // Process parsed addresses.
-    for (const a of parsed) {
-      a.isImported && !isOnline && (a.isImported = false);
-
-      // Persist or update address in Electron store.
-      await window.myAPI.rawAccountTask({
-        action: 'raw-account:import',
-        data: { source, serialized: JSON.stringify(a) },
-      });
-
-      // Add address and its status to import window's state.
-      importWindowOpen &&
-        postToImport('import:account:add', { json: JSON.stringify(a), source });
-
-      // Handle importing or removing account from main window and setting `isImported` flag state.
-      const { address, name } = a;
-      const chainId = getAddressChainId(address);
-
-      if (a.isImported) {
-        const data = { data: { data: { address, chainId, name, source } } };
-        await handleImportAddress(new MessageEvent('message', data), true);
-        postToImport('import:address:update', { address: a, source });
-      } else {
-        const data = { data: { data: { address, chainId } } };
-        await handleRemoveAddress(new MessageEvent('message', data));
-        postToImport('import:address:update', { address: a, source });
-      }
-
-      // Update managed account names.
-      const account = AccountsController.get(chainId, address);
-      if (account) {
-        account.name = name;
-        AccountsController.update(chainId, account);
-      }
-    }
-
-    // Update account list state.
-    setAddresses(AccountsController.getAllFlattenedAccountData());
-  }
-};
-
-/**
- * @name importEvents
- * @summary Extract event data from an imported text file and send to application.
- * (main renderer)
- */
-export const importEvents = async (
-  serialized: string,
-  setEvents: (events: EventCallback[]) => void
-): Promise<void> => {
-  const s_events = getFromBackupFile('events', serialized);
-  if (!s_events) {
-    return;
-  }
-
-  // Send serialized events to main for processing.
-  const updated = (await window.myAPI.sendEventTaskAsync({
-    action: 'events:import',
-    data: { events: s_events },
-  })) as string;
-
-  const parsed: EventCallback[] = JSON.parse(updated);
-  setEvents(parsed);
-};
-
-/**
- * @name importIntervalTasks
- * @summary Extract interval task data from an imported text file and send to application.
- * (main renderer)
- */
-type IntervalFunc = (t: IntervalSubscription) => void;
-
-export const importIntervalTasks = async (
-  serialized: string,
-  tryAddIntervalSubscription: IntervalFunc,
-  tryUpdateDynamicIntervalTask: IntervalFunc,
-  addIntervalSubscription: IntervalFunc,
-  updateIntervalSubscription: IntervalFunc
-): Promise<void> => {
-  const s_tasks = getFromBackupFile('intervals', serialized);
-  if (!s_tasks) {
-    return;
-  }
-
-  // Receive new tasks after persisting them to store.
-  const s_data =
-    (await window.myAPI.sendIntervalTask({
-      action: 'interval:tasks:import',
-      data: { serialized: s_tasks },
-    })) || '[]';
-
-  // Parse received tasks to insert and update.
-  const s_array: [string, string][] = JSON.parse(s_data);
-  const map = new Map<string, string>(s_array);
-
-  const inserts: IntervalSubscription[] = JSON.parse(map.get('insert') || '[]');
-  const updates: IntervalSubscription[] = JSON.parse(map.get('update') || '[]');
-
-  // Update manage subscriptions in controller and update React state.
-  if (inserts.length > 0) {
-    IntervalsController.insertSubscriptions(inserts);
-    inserts.forEach((t) => {
-      tryAddIntervalSubscription(t);
-      addIntervalSubscription(t);
-    });
-  }
-
-  if (updates.length > 0) {
-    IntervalsController.removeSubscriptions(updates);
-    updates.forEach((t) => {
-      t.status === 'enable' && IntervalsController.insertSubscription(t);
-      tryUpdateDynamicIntervalTask(t);
-      updateIntervalSubscription(t);
-    });
-  }
-
-  // Update state in OpenGov window.
-  if (await window.myAPI.isViewOpen('openGov')) {
-    inserts.forEach((t) => {
-      ConfigRenderer.portToOpenGov?.postMessage({
-        task: 'openGov:task:add',
-        data: {
-          serialized: JSON.stringify(t),
-        },
-      });
-    });
-
-    updates.forEach((t) => {
-      ConfigRenderer.portToOpenGov?.postMessage({
-        task: 'openGov:task:update',
-        data: {
-          serialized: JSON.stringify(t),
-        },
-      });
-    });
-  }
-};
-
-/**
- * @name importAccountSubscriptions
- * @summary Extract account subscription data from an imported text file and send to application.
- */
-export const importAccountSubscriptions = async (
-  serialized: string,
-  updateRenderedSubscriptions: (task: SubscriptionTask) => void,
-  setAccountSubscriptions: (m: Map<string, SubscriptionTask[]>) => void
-): Promise<void> => {
-  const s_tasks = getFromBackupFile('accountTasks', serialized);
-  if (!s_tasks) {
-    return;
-  }
-
-  const s_array: [string, string][] = JSON.parse(s_tasks);
-  const s_map = new Map<string, string>(s_array);
-
-  // Store tasks to persist to store.
-  const s_persistMap = new Map<string, string>();
-
-  // Iterate map of serialized tasks keyed by an account address.
-  for (const [address, serTasks] of s_map.entries()) {
-    const parsed: SubscriptionTask[] = JSON.parse(serTasks);
-    if (parsed.length === 0) {
-      continue;
-    }
-
-    const account = AccountsController.get(parsed[0].chainId, address);
-    const valid: SubscriptionTask[] = [];
-
-    if (account) {
-      for (const t of parsed) {
-        if (
-          (t.category === 'Nomination Pools' && !account.nominationPoolData) ||
-          (t.category === 'Nominating' && !account.nominatingData)
-        ) {
-          // Throw away task if necessary.
-          continue;
-        }
-
-        // Otherwise subscribe to task.
-        await account?.subscribeToTask(t);
-        updateRenderedSubscriptions(t);
-        valid.push(t);
-      }
-    }
-
-    // Serialize the account's subscribed tasks.
-    valid.length > 0 && s_persistMap.set(address, JSON.stringify(valid));
-  }
-
-  // Set subscriptions React state.
-  setAccountSubscriptions(
-    SubscriptionsController.getAccountSubscriptions(AccountsController.accounts)
-  );
-
-  // Send successfully imported tasks to main process.
-  await window.myAPI.sendSubscriptionTask({
-    action: 'subscriptions:account:import',
-    data: { serialized: JSON.stringify(Array.from(s_persistMap.entries())) },
-  });
-};
-
-/**
  * @name getFromBackupFile
  * @summary Get some serialized data from backup files.
  * Key may be `addresses`, `events` or `intervals`.
  */
-const getFromBackupFile = (
+export const getFromBackupFile = (
   key: string,
   serialized: string
 ): string | undefined => {
@@ -421,6 +178,24 @@ const getFromBackupFile = (
  * @summary Utility to post a message to the import window.
  * (main renderer)
  */
-const postToImport = (task: string, dataObj: AnyData) => {
+export const postToImport = (task: string, dataObj: AnyData) => {
   ConfigRenderer.portToImport?.postMessage({ task, data: dataObj });
+};
+
+/**
+ * @name postToOpenGov
+ * @summary Utility to post a message to the OpenGov window.
+ * (main renderer)
+ */
+export const postToOpenGov = (task: string, dataObj: AnyData) => {
+  ConfigRenderer.portToOpenGov?.postMessage({ task, data: dataObj });
+};
+
+/**
+ * @name postToSettings
+ * @summary Utility to post a message to the Settings window.
+ * (main renderer)
+ */
+export const postToSettings = (task: string, dataObj: AnyData) => {
+  ConfigRenderer.portToSettings?.postMessage({ task, data: dataObj });
 };
